feat: update svg images and move index.html to the root project
