+++ package.json
@@ -86,7 +86,7 @@
   "scripts": {
     "preinstall": "node scripts/use-yarn.js",
     "prepare": "husky install",
-    "dev": "craco start",
+    "dev": "vite",
     "start": "craco start",
     "build": "craco build",
     "vercel:install": "yarn run add:npmrc && yarn install",
