+++ src/app/auth/views/RecoverAccountView/RecoverAccountView.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import { Button } from '@internxt/ui';
 import { Question } from '@phosphor-icons/react';
-import { ReactComponent as InternxtLogo } from 'assets/icons/big-logo.svg';
+import InternxtLogo from 'assets/icons/big-logo.svg?react';
 import ChangePassword from 'app/auth/components/ChangePassword/ChangePassword';
 import RestartAccount from 'app/auth/components/RestartAccount/RestartAccount';
