+++ src/app/auth/views/RecoveryLinkView/RecoveryLinkView.tsx
@@ -1,5 +1,5 @@
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
-import { ReactComponent as InternxtLogo } from 'assets/icons/big-logo.svg';
+import InternxtLogo from 'assets/icons/big-logo.svg?react';
 import RecoveryLink from '../../components/RecoveryLink/RecoveryLink';
 
 function RecoveryLinkView(): JSX.Element {
