+++ src/app/core/services/env.service.node.test.ts
@@ -56,8 +56,6 @@ describe('Check that env variables are loaded correctly', () => {
     expect(envConfig.analytics.cdpDataPlane).toBe(process.env.REACT_APP_CDP_DATA_PLANE);
 
     expect(envConfig.app.nodeEnv).toBe(process.env.NODE_ENV);
-    expect(envConfig.app.fastRefresh).toBe(process.env.FAST_REFRESH);
-    expect(envConfig.app.debug).toBe(process.env.REACT_APP_DEBUG);
     expect(envConfig.app.generateSourceMap).toBe(process.env.GENERATE_SOURCEMAP);
     expect(envConfig.app.hostname).toBe(process.env.REACT_APP_HOSTNAME);
     expect(envConfig.app.websiteUrl).toBe(process.env.REACT_APP_WEBSITE_URL);
