+++ src/app/core/services/env.service.ts
@@ -1,6 +1,6 @@
 function isProduction(): boolean {
-  if (process.env.REACT_APP_NODE_ENV === 'staging') return false;
-  return process.env.NODE_ENV === 'production' || process.env.REACT_APP_NODE_ENV === 'production';
+  if (import.meta.env.VITE_NODE_ENV === 'staging') return false;
+  return import.meta.env.NODE_ENV === 'production' || import.meta.env.VITE_NODE_ENV === 'production';
 }
 
 const envService = {
@@ -10,15 +10,14 @@ const envService = {
 export default envService;
 
 function getEnvVar(name: keyof NodeJS.ProcessEnv): string {
-  const value = process.env[name];
+  const value = import.meta.env[name];
   return value!;
 }
 
 export const envConfig = {
   app: {
     nodeEnv: getEnvVar('NODE_ENV'),
-    fastRefresh: getEnvVar('FAST_REFRESH'),
-    debug: getEnvVar('REACT_APP_DEBUG'),
+    debug: getEnvVar('VITE_DEBUG'),
     generateSourceMap: getEnvVar('GENERATE_SOURCEMAP'),
     hostname: getEnvVar('REACT_APP_HOSTNAME'),
     websiteUrl: getEnvVar('REACT_APP_WEBSITE_URL'),
