+++ src/app/drive/views/RequestAccess/RequestAccess.tsx
@@ -5,7 +5,7 @@ import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
 import { LockSimple, CheckCircle } from '@phosphor-icons/react';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import { Button } from '@internxt/ui';
-import { ReactComponent as InternxtLogo } from 'assets/icons/big-logo.svg';
+import InternxtLogo from 'assets/icons/big-logo.svg?react';
 import './RequestAccess.scss';
 import { logOut } from '../../../auth/services/auth.service';
